Make ProjectCreateTests tests as non-api tests
